s10 - eos - Handling errors REACT/.NET

+++ Application/Activities/Delete.cs
@@ -1,3 +1,5 @@
+using System.Data.Common;
+using Application.Core;
 using Domain;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
@@ -8,12 +10,12 @@ namespace Application.Activities
     public class Delete
     {
         // Parametros a serem recebidos para a criação da atividade
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
 
@@ -22,13 +24,20 @@ namespace Application.Activities
                 _context = context;
             }
 
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await _context.Activities.FindAsync(request.Id);
 
+                if (activity == null) return null;
+
                 _context.Activities.Remove(activity);
 
-                await _context.SaveChangesAsync();
+                var result = await _context.SaveChangesAsync() > 0;
+
+                if (!result) return Result<Unit>.Failure("Falha ao deletar atividade");
+
+                return Result<Unit>.Success(Unit.Value);
+
             }
         }
     }
